requirement.txt
absl-py==0.9.0
astor==0.8.1
attrs==19.3.0
autobahn==20.3.1
Automat==20.2.0
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
cycler==0.10.0
dateparser==0.7.4
gast==0.2.2
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.27.2
h5py==2.10.0
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.2.1
matplotlib==3.2.0
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.2.0
pandas==1.0.1
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.6
python-binance==0.7.5
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
regex==2020.2.20
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
service-identity==18.1.0
six==1.14.0
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
Twisted==19.10.0
txaio==20.1.1
tzlocal==2.0.0
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.12.1
zope.interface==4.7.2
